third_party/blink/web_tests/http/tests/misc/resource-timing-navigation-in-restored-iframe-same-to-cross-expected.txt

ALERT: Going back.
Tests that subsequent navigation in an iframe restored from history does not report resource timing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS resources.length is 3
PASS resources[0].name is "http://127.0.0.1:8000/js-test-resources/js-test.js"
PASS resources[1].name is "http://127.0.0.1:8000/resources/prevent-bfcache.js"
PASS resources[2].name is "http://127.0.0.1:8000/misc/resources/navigate-on-message.html"
PASS successfullyParsed is true

TEST COMPLETE


Back-forward cache keeps and restores the resource timing of the iframe on back navigation. Back-forward cache needs to be disabled to run this test.
